+++ .editorconfig
@@ -11,7 +11,7 @@ insert_final_newline = true
 
 # Matches multiple files with brace expansion notation
 # Set default charset
-[*.{js,json}]
+[*.{js,jsx,json}]
 indent_size = 2
 indent_style = space
