+++ src/components/App.js
@@ -1,44 +1,18 @@
 import React from 'react';
-import { withStyles } from 'material-ui/styles';
-import { MuiThemeProvider, createMuiTheme } from 'material-ui/styles';
-import Sidebar from './Sidebar.jsx';
-import ChatHeader from './ChatHeader.jsx';
-import Chat from './Chat.jsx';
-import { chats, messages } from '../data.json';
-import red from 'material-ui/colors/red';
-import grey from 'material-ui/colors/grey';
+import { BrowserRouter as Router, Route, Link, Switch, Redirect } from 'react-router-dom';
+import WelcomePage from './WelcomePage.jsx'
+import ChatPage from './ChatPage.jsx';
 
-const theme = createMuiTheme({
-  palette: {
-    primary: {
-      light: red[200],
-      main: red[500],
-      dark: red[700],
-      contrastText: grey[50],
-    }
-  }
-});
 
-const styles = theme => ({
-  root: {
-    position: 'relative',
-    display: 'flex',
-    width: '100%',
-    height: '100%',
-    zIndex: 1,
-    overflow: 'hidden',
-    backgroundColor: theme.palette.background.default
-  }
-});
-
-const App = ({ classes }) => ( 
-  <MuiThemeProvider theme={theme}>
-    <div className={classes.root}>       
-      <ChatHeader />
-      <Sidebar chats={chats}/>
-      <Chat messages={messages}/>
-    </div>
-  </MuiThemeProvider>
+const App = () => (
+  <Router>
+    <Switch>
+      <Route exact path='/(welcome)?' component={WelcomePage} />
+      <Route path='/chat' component={ChatPage} />
+      <Redirect to="/" />
+    </Switch>
+  </Router>
 );
 
-export default withStyles(styles)(App);
+
+export default App;
