+++ src/components/WelcomePage.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class WelcomePage extends Component {
+  render() {
+    return (
+      <div>
+        <h1>Welcome</h1>
+      </div>
+    );
+  }
+}
+
+export default WelcomePage;
